# Optimize Encoder.WriteValue for object names